ItemSelected to putExtra & Fixed ArrayAdapter Disp
For starting the new activity, I included the selected item in the
putExtra so that we don't have to rely on the content of the View to
determine which position was selected.

ArrayAdapter Disp Fix
Array adapter always displayed the wrong list details when scrolling
past the bottom item on screen or scrolling back up.  This has been
fixed.